packages: Add International Components for Unicode fetching
This has been added as on FreeBSD platform it has to be build from source.
A version available in pkg repository uses libcxxrt ABI library, whereas
rest of the agent is build with libsubc++ that are not compatible with
each other and are causing the Agent to crash.

Change-Id: I32f51604790331ee61df6b90257af5c4f7a259a1
(cherry picked from commit 4dae1102a236e39b09c7753d7d65e80f25753226)